Moved CompanyCreate* components to their own folder
